+++ sdk/storage/azure-storage-file-share/azure/storage/fileshare/aio/_share_service_client_async.py
@@ -22,6 +22,7 @@ from .._generated.aio import AzureFileStorage
 from .._generated.models import StorageErrorException, StorageServiceProperties
 from .._generated.version import VERSION
 from .._share_service_client import ShareServiceClient as ShareServiceClientBase
+from .._serialize import get_api_version
 from ._share_client_async import ShareClient
 from ._models import SharePropertiesPaged
 from .._models import service_properties_deserialize
@@ -55,6 +56,9 @@ class ShareServiceClient(AsyncStorageAccountHostsMixin, ShareServiceClientBase):
     :keyword str api_version:
         The Storage API version to use for requests. Default value is '2019-07-07'.
         Setting to an older version may result in reduced feature compatibility.
+
+        .. versionadded:: 12.1.0
+
     :keyword str secondary_hostname:
         The hostname of the secondary endpoint.
     :keyword loop:
@@ -84,7 +88,7 @@ class ShareServiceClient(AsyncStorageAccountHostsMixin, ShareServiceClientBase):
             loop=loop,
             **kwargs)
         self._client = AzureFileStorage(version=VERSION, url=self.url, pipeline=self._pipeline, loop=loop)
-        self._client._config.version = kwargs.get('api_version', VERSION)  # pylint: disable=protected-access
+        self._client._config.version = get_api_version(kwargs, VERSION)  # pylint: disable=protected-access
         self._loop = loop
 
     @distributed_trace_async